Repo

+++ Repo/Controllers/BaseController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
 using Repo.Classes;
 using Repo.Interfaces;
@@ -11,56 +12,107 @@ using Repo.Interfaces;
 namespace Repo.Controllers
 {
     [Route("api/[controller]")]
-    public class BaseController<TEntity> : Controller where TEntity : class
+    public class BaseController<TEntity, TDtoGet, TDtoPost, TDtoPut> : Controller where TEntity : class where TDtoGet : class where TDtoPost : class where TDtoPut : class
     {
         private IRepository<TEntity> _repository;
         private IUnitOfWork _unitOfWork;
+        private IMapper _mapper;
 
-        public BaseController(IRepository<TEntity> repository, IUnitOfWork unitOfWork)
+        public BaseController(IRepository<TEntity> repository, IUnitOfWork unitOfWork, IMapper mapper)
         {
             _repository = repository;
             _unitOfWork = unitOfWork;
+            _mapper = mapper;
         }
 
         // GET: api/<controller>
         [HttpGet]
         public IActionResult Get()
         {
-            return Ok(_repository.Get());
+            var data = _repository.Get();
+
+            if (data != null)
+            {
+                var otp = _mapper.Map<IEnumerable<TDtoGet>>(data);
+
+                return Ok(otp);
+            }
+
+            return NotFound();
         }
 
         // GET api/<controller>/5
         [HttpGet("{id}")]
         public IActionResult Get(int id)
         {
-            return Ok(_repository.GetById(id));
+            var data = _repository.GetById(id);
+
+            if (data != null)
+            {
+                var otp = _mapper.Map<TDtoGet>(data);
+
+                return Ok(otp);
+            }
+
+            return NotFound();
         }
 
         // POST api/<controller>
         [HttpPost]
-        public IActionResult Post([FromBody]TEntity entity)
+        public IActionResult Post(TDtoPost input)
         {
-            _repository.Add(entity);
-            _unitOfWork.Save();
-            return Ok();
+            try
+            {
+                var otp = _mapper.Map<TEntity>(input);
+
+                _repository.Add(otp);
+                _unitOfWork.Save();
+                return Ok();
+            }
+            catch (Exception)
+            {
+                return BadRequest();
+            }
         }
 
         // PUT api/<controller>/5
         [HttpPut("{id}")]
-        public IActionResult Put(int id, [FromBody]TEntity entity)
+        public IActionResult Put(int id, TDtoPut input)
         {
-            _repository.Update(entity);
-            _unitOfWork.Save();
-            return Ok();
+            try
+            {
+                var entity = _repository.GetById(id);
+                //_repository.Update(entity);
+                _mapper.Map(input, entity);
+                _unitOfWork.Save();
+                return Ok();
+
+            }
+            catch (Exception)
+            {
+                return BadRequest();
+            }
         }
 
         // DELETE api/<controller>/5
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
-            _repository.Remove(id);
-            _unitOfWork.Save();
-            return Ok();
+            if (id != 0)
+            {
+                try
+                {
+                    _repository.Remove(id);
+                    _unitOfWork.Save();
+                    return Ok();
+                }
+                catch (Exception)
+                {
+                    return BadRequest();
+                }
+            }
+
+            return NotFound();
         }
     }
 }
